on people admin page loads, verify the default term for people is set, if not, then set it. [touch: 7780]
Note: I also removed the setting of a person to the default term on save.  This is because it's kind of buggy and I'd rather have users explicitly set the term then have unexpected results.

The verification also does NOT recreate a staff term if that term was already created and then later edited to have a different slug.